# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express MySQL Docker API",
    "version": "1.0.0",
    "description": "ノート・クイズ管理アプリケーションの REST API"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API サーバー"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT トークンを Authorization ヘッダーに設定"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "全ユーザー一覧取得",
        "description": "全ユーザーの一覧を取得します。Bearer認証および管理者（admin）ロールが必要です。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "取得成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "ユーザー登録",
        "description": "新規ユーザーを登録します。認証不要（会員登録用）。ロールはデフォルトで\"user\"になります。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "ユーザー名"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "メールアドレス"
                  },
                  "password": {
                    "type": "string",
                    "description": "パスワード"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "登録成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "role": {
                      "type": "string",
                      "example": "user",
                      "emailVerified": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "必須パラメータ不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Name, email, and password are required"
                }
              }
            }
          },
          "409": {
            "description": "ユーザーが既に存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "User already exists"
                }
              }
            }
          }
        }
      }
    }
  },
...
}